# Refactoring: implement arrays for traceql.Static with reused fields