--- a/pkg/traceql/ast_conditions.go
+++ b/pkg/traceql/ast_conditions.go
@@ -13,7 +13,12 @@ func (f SpansetFilter) extractConditions(request *FetchSpansRequest) {
 	// For empty spansets { } ensure there is something that matches all spans.
 	// Use start time which would have been selected as part of the second pass
 	// metadata, and is still fairly efficient to pull back.
-	if s, ok := f.Expression.(Static); ok && s.Type == TypeBoolean && s.B {
+	if s, ok := f.Expression.(Static); ok {
+		b, err := s.Bool()
+		if err != nil || !b {
+			return
+		}
+
 		for _, c := range request.Conditions {
 			if c.Attribute.Intrinsic != IntrinsicNone && c.Op == OpNone {
 				// A different match-all intrinsic is already present.