… CHANGELOG.md
… pkg/traceql/ast_execute.go
… pkg/traceql/ast_stringer.go
… pkg/traceql/ast_test.go
… pkg/traceql/ast_validate.go
… pkg/traceql/engine.go
… pkg/traceql/engine_metrics.go
… pkg/traceql/engine_metrics_compare.go
… pkg/traceql/enum_hints.go
… pkg/traceql/enum_statics.go
… pkg/traceql/util.go
… pkg/traceqlmetrics/metrics.go
… pkg/traceqlmetrics/metrics_test.go
… tempodb/encoding/vparquet2/block_traceql.go
… tempodb/encoding/vparquet3/block_autocomplete.go
… tempodb/encoding/vparquet3/block_traceql.go
… tempodb/encoding/vparquet3/coalesce_conditions.go
… tempodb/encoding/vparquet4/block_autocomplete.go
… tempodb/encoding/vparquet4/block_traceql.go
… tempodb/encoding/vparquet4/coalesce_conditions.go
… tempodb/tempodb_search_test.go