**What this PR does**:
This PR adds support for the new types `TypeIntArray`, `TypeFloatArray`, `TypeStringArray`, and `TypeBooleanArray` to  `traceql.Static`. This is a perquisite for supporting arrays in TraceQL.

The goal of this approach is to add support for array types to `traceql.Static` without increasing the size of the existing struct. It makes use of packed/reused struct fields that store values of multiple different types. This is accomplished through heavy use of casts and `unsafe`.

The new `traceql.Static` type is no longer comparable and can no longer be used as key in a map. This PR therefore introduces a new type `traceql.StaticMapKey`.

For context: a previous approach based on a `traceql.Static` interface with multiple implementations was [explored in this PR](https://github.com/grafana/tempo/pull/3783).

**Which issue(s) this PR fixes**:
Is prerequisite for https://github.com/grafana/tempo-squad/issues/435

**Checklist**
- [x] Tests updated
- [ ] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`