@stoewer 2024-07-02
The first benchmark results for this approach are quite promising:

Metrics:
```
goos: linux
goarch: amd64
pkg: github.com/grafana/tempo/tempodb/encoding/vparquet4
cpu: AMD Ryzen 7 PRO 6850U with Radeon Graphics     
                                                                             │ bench-metrics-02-static.txt │ bench-metrics-02-static-reuse-fields.txt │
                                                                             │           sec/op            │      sec/op        vs base               │
BackendBlockQueryRange/{}_|_rate()/7-16                                                        2.361 ± 12%         2.445 ± 10%       ~ (p=0.218 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-16                                              4.508 ± 23%         4.541 ± 18%       ~ (p=0.912 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-16                             3.520 ± 14%         3.528 ± 14%       ~ (p=0.971 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-16                                     4.260 ±  3%         4.307 ±  7%       ~ (p=0.436 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-16                  236.8m ±  5%        236.7m ±  6%       ~ (p=0.971 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/7-16                                           290.1m ±  0%        291.3m ±  1%       ~ (p=0.063 n=10)
geomean                                                                                        1.490               1.505        +1.00%

                                                                             │ bench-metrics-02-static.txt │ bench-metrics-02-static-reuse-fields.txt │
                                                                             │          MB_IO/op           │    MB_IO/op      vs base                 │
BackendBlockQueryRange/{}_|_rate()/7-16                                                         34.39 ± 0%        34.39 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-16                                               44.62 ± 0%        44.62 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-16                              34.87 ± 0%        34.87 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-16                                      38.73 ± 0%        38.73 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-16                    34.87 ± 0%        34.87 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{status=error}_|_rate()/7-16                                             36.07 ± 0%        36.07 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                                         37.10             37.10       +0.00%
¹ all samples are equal

                                                                             │ bench-metrics-02-static.txt │ bench-metrics-02-static-reuse-fields.txt │
                                                                             │          spans/op           │    spans/op      vs base                 │
BackendBlockQueryRange/{}_|_rate()/7-16                                                        8.537M ± 0%       8.537M ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-16                                              8.537M ± 0%       8.537M ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-16                             8.537M ± 0%       8.537M ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-16                                     8.537M ± 0%       8.537M ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-16                   382.0k ± 0%       382.0k ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{status=error}_|_rate()/7-16                                            94.10k ± 0%       94.10k ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                                        2.400M            2.400M       +0.00%
¹ all samples are equal

                                                                             │ bench-metrics-02-static.txt │ bench-metrics-02-static-reuse-fields.txt │
                                                                             │           spans/s           │      spans/s       vs base               │
BackendBlockQueryRange/{}_|_rate()/7-16                                                       3.616M ± 13%        3.492M ± 11%       ~ (p=0.218 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-16                                             1.894M ± 29%        1.880M ± 22%       ~ (p=0.912 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-16                            2.426M ± 16%        2.420M ± 17%       ~ (p=0.971 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-16                                    2.004M ±  3%        1.982M ±  7%       ~ (p=0.436 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-16                  1.613M ±  5%        1.614M ±  7%       ~ (p=0.971 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/7-16                                           324.4k ±  0%        323.1k ±  1%       ~ (p=0.063 n=10)
geomean                                                                                       1.610M              1.594M        -0.99%

                                                                             │ bench-metrics-02-static.txt │ bench-metrics-02-static-reuse-fields.txt │
                                                                             │            B/op             │       B/op         vs base               │
BackendBlockQueryRange/{}_|_rate()/7-16                                                      51.50Mi ± 29%       50.28Mi ± 29%       ~ (p=0.424 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-16                                            72.85Mi ± 23%       70.19Mi ± 31%       ~ (p=0.105 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-16                           58.76Mi ± 56%       59.25Mi ± 37%       ~ (p=0.971 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-16                                   443.9Mi ±  7%       433.2Mi ±  9%       ~ (p=0.393 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-16                 16.16Mi ± 13%       15.21Mi ± 22%       ~ (p=0.353 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/7-16                                          5.788Mi ± 23%       6.151Mi ± 21%       ~ (p=0.436 n=10)
geomean                                                                                      45.74Mi             45.15Mi        -1.27%

                                                                             │ bench-metrics-02-static.txt │ bench-metrics-02-static-reuse-fields.txt │
                                                                             │          allocs/op          │     allocs/op      vs base               │
BackendBlockQueryRange/{}_|_rate()/7-16                                                        522.6k ± 0%        522.6k ±  0%       ~ (p=0.104 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-16                                              540.2k ± 0%        540.2k ±  0%       ~ (p=0.100 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-16                             523.9k ± 0%        523.9k ±  0%       ~ (p=0.926 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-16                                     2.102M ± 8%        2.024M ± 12%       ~ (p=0.393 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-16                   117.8k ± 0%        117.8k ±  0%       ~ (p=0.492 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/7-16                                            76.53k ± 0%        76.53k ±  0%       ~ (p=0.323 n=10)
geomean     
``` 

TraceQL
```
goos: linux
goarch: amd64
pkg: github.com/grafana/tempo/tempodb/encoding/vparquet4
cpu: AMD Ryzen 7 PRO 6850U with Radeon Graphics     
                                                   │ bench-traceql-02-static.txt │ bench-traceql-02-static-reuse-fields.txt │
                                                   │           sec/op            │      sec/op       vs base                │
BackendBlockTraceQL/spanAttValMatch-16                               554.7m ± 0%        549.8m ± 1%   -0.87% (p=0.007 n=10)
BackendBlockTraceQL/spanAttValNoMatch-16                             75.43m ± 1%        78.58m ± 3%   +4.17% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-16                         165.8m ± 0%        174.4m ± 1%   +5.24% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                       4.194m ± 0%        4.356m ± 1%   +3.87% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValMatch-16                            1.830 ± 1%         1.851 ± 1%   +1.18% (p=0.011 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-16                         3.802m ± 1%        3.936m ± 3%   +3.53% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16                     325.1m ± 0%        363.2m ± 1%  +11.70% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-16                   3.864m ± 1%        4.066m ± 0%   +5.20% (p=0.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-16                               665.0m ± 1%        681.4m ± 1%   +2.47% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                         53.85m ± 1%        55.53m ± 1%   +3.12% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-16                          548.8m ± 1%        555.8m ± 2%   +1.28% (p=0.002 n=10)
BackendBlockTraceQL/count-16                                          1.466 ± 1%         1.587 ± 2%   +8.23% (p=0.000 n=10)
BackendBlockTraceQL/struct-16                                         2.468 ± 4%         2.458 ± 3%        ~ (p=0.853 n=10)
BackendBlockTraceQL/||-16                                            558.4m ± 1%        589.2m ± 1%   +5.52% (p=0.000 n=10)
BackendBlockTraceQL/mixed-16                                         205.4m ± 1%        221.2m ± 1%   +7.71% (p=0.000 n=10)
BackendBlockTraceQL/complex-16                                       4.343m ± 0%        4.671m ± 1%   +7.54% (p=0.000 n=10)
geomean                                                              132.5m             138.2m        +4.29%

                                                   │ bench-traceql-02-static.txt │ bench-traceql-02-static-reuse-fields.txt │
                                                   │             B/s             │       B/s         vs base                │
BackendBlockTraceQL/spanAttValMatch-16                              107.2Mi ± 0%       108.2Mi ± 1%   +0.88% (p=0.007 n=10)
BackendBlockTraceQL/spanAttValNoMatch-16                            235.0Mi ± 1%       225.6Mi ± 3%   -4.00% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-16                        375.2Mi ± 0%       356.6Mi ± 1%   -4.98% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                      649.9Mi ± 0%       625.7Mi ± 1%   -3.73% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValMatch-16                          31.86Mi ± 1%       31.49Mi ± 1%   -1.17% (p=0.011 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-16                        197.0Mi ± 1%       190.3Mi ± 3%   -3.41% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16                    179.8Mi ± 0%       160.9Mi ± 1%  -10.48% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-16                  229.6Mi ± 1%       218.2Mi ± 0%   -4.95% (p=0.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-16                              31.42Mi ± 1%       30.67Mi ± 1%   -2.40% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                        475.1Mi ± 1%       460.8Mi ± 1%   -3.02% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-16                         48.81Mi ± 1%       48.20Mi ± 2%   -1.26% (p=0.002 n=10)
BackendBlockTraceQL/count-16                                        39.76Mi ± 1%       36.74Mi ± 2%   -7.60% (p=0.000 n=10)
BackendBlockTraceQL/struct-16                                       5.522Mi ± 4%       5.541Mi ± 3%        ~ (p=0.869 n=10)
BackendBlockTraceQL/||-16                                          104.96Mi ± 1%       99.47Mi ± 1%   -5.23% (p=0.000 n=10)
BackendBlockTraceQL/mixed-16                                        282.0Mi ± 1%       261.8Mi ± 1%   -7.16% (p=0.000 n=10)
BackendBlockTraceQL/complex-16                                      735.5Mi ± 0%       683.3Mi ± 1%   -7.10% (p=0.000 n=10)
geomean                                                             128.5Mi            123.2Mi        -4.12%

                                                   │ bench-traceql-02-static.txt │ bench-traceql-02-static-reuse-fields.txt │
                                                   │          MB_io/op           │    MB_io/op      vs base                 │
BackendBlockTraceQL/spanAttValMatch-16                                62.38 ± 0%        62.38 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttValNoMatch-16                              18.59 ± 0%        18.59 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttIntrinsicMatch-16                          65.22 ± 0%        65.22 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                        2.858 ± 0%        2.858 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttValMatch-16                            61.14 ± 0%        61.14 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttValNoMatch-16                         785.4m ± 0%       785.4m ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttIntrinsicMatch-16                      61.29 ± 0%        61.29 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-16                   930.2m ± 0%       930.2m ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValNoMatch-16                                21.91 ± 0%        21.91 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValMixedMatchAnd-16                          26.83 ± 0%        26.83 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValMixedMatchOr-16                           28.09 ± 0%        28.09 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/count-16                                          61.12 ± 0%        61.12 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/struct-16                                         14.29 ± 0%        14.29 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/||-16                                             61.46 ± 0%        61.46 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixed-16                                          60.73 ± 0%        60.73 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/complex-16                                        3.351 ± 0%        3.352 ± 0%       ~ (p=0.739 n=10)
geomean                                                               17.86             17.86       +0.00%
¹ all samples are equal

                                                   │ bench-traceql-02-static.txt │ bench-traceql-02-static-reuse-fields.txt │
                                                   │            B/op             │       B/op         vs base               │
BackendBlockTraceQL/spanAttValMatch-16                             250.2Mi ±  0%       236.3Mi ±  1%  -5.54% (p=0.000 n=10)
BackendBlockTraceQL/spanAttValNoMatch-16                           29.37Mi ±  1%       29.42Mi ±  0%       ~ (p=0.218 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-16                       29.55Mi ±  3%       29.47Mi ±  2%       ~ (p=0.796 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                     2.734Mi ±  0%       2.740Mi ±  0%       ~ (p=0.218 n=10)
BackendBlockTraceQL/resourceAttValMatch-16                         1.547Gi ±  0%       1.536Gi ±  0%  -0.69% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-16                       2.667Mi ±  0%       2.670Mi ±  1%       ~ (p=0.280 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16                   145.5Mi ±  1%       138.2Mi ±  1%  -5.02% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-16                 2.672Mi ±  0%       2.669Mi ±  0%       ~ (p=0.529 n=10)
BackendBlockTraceQL/mixedValNoMatch-16                             31.50Mi ± 10%       32.41Mi ±  3%       ~ (p=0.247 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                       3.614Mi ±  7%       3.549Mi ± 11%       ~ (p=0.853 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-16                        4.838Mi ± 15%       4.593Mi ± 27%       ~ (p=0.089 n=10)
BackendBlockTraceQL/count-16                                       1.129Gi ±  0%       1.117Gi ±  1%  -1.03% (p=0.000 n=10)
BackendBlockTraceQL/struct-16                                      536.0Mi ±  1%       539.2Mi ±  1%       ~ (p=0.063 n=10)
BackendBlockTraceQL/||-16                                          19.80Mi ±  7%       18.98Mi ±  3%       ~ (p=0.105 n=10)
BackendBlockTraceQL/mixed-16                                       22.75Mi ±  4%       21.78Mi ±  3%  -4.28% (p=0.015 n=10)
BackendBlockTraceQL/complex-16                                     2.792Mi ±  1%       2.780Mi ±  0%       ~ (p=0.165 n=10)
geomean                                                            29.95Mi             29.48Mi        -1.56%

                                                   │ bench-traceql-02-static.txt │ bench-traceql-02-static-reuse-fields.txt │
                                                   │          allocs/op          │     allocs/op      vs base               │
BackendBlockTraceQL/spanAttValMatch-16                               2.174M ± 0%         2.174M ± 0%  -0.00% (p=0.000 n=10)
BackendBlockTraceQL/spanAttValNoMatch-16                             42.03k ± 0%         42.03k ± 0%       ~ (p=0.971 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-16                         207.8k ± 0%         207.8k ± 0%       ~ (p=0.209 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                       41.69k ± 0%         41.69k ± 0%       ~ (p=0.628 n=10)
BackendBlockTraceQL/resourceAttValMatch-16                           10.04M ± 0%         10.04M ± 0%       ~ (p=0.149 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-16                         41.72k ± 0%         41.72k ± 0%       ~ (p=0.087 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16                     1.360M ± 0%         1.360M ± 0%       ~ (p=0.170 n=10)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-16                   41.72k ± 0%         41.72k ± 0%       ~ (p=1.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-16                               46.69k ± 0%         46.70k ± 0%       ~ (p=0.288 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                         43.72k ± 0%         43.72k ± 0%       ~ (p=0.687 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-16                          44.71k ± 0%         44.71k ± 0%       ~ (p=0.147 n=10)
BackendBlockTraceQL/count-16                                         5.982M ± 0%         5.982M ± 0%       ~ (p=0.956 n=10)
BackendBlockTraceQL/struct-16                                        4.386M ± 0%         4.395M ± 0%  +0.21% (p=0.015 n=10)
BackendBlockTraceQL/||-16                                            207.1k ± 0%         207.1k ± 0%       ~ (p=0.342 n=10)
…